Extract datasetpageheader into sepparate file
